Clean/everything

+++ src/Weblu.Application/Services/ServiceServices/ServiceService.cs
@@ -1,14 +1,15 @@
 using AutoMapper;
-using Weblu.Application.Dtos.ServiceDtos;
-using Weblu.Application.Dtos.ServiceDtos.ServiceImageDtos;
-using Weblu.Application.Exceptions;
 using Weblu.Domain.Entities.Services;
 using Weblu.Domain.Errors.Services;
-using Weblu.Application.Interfaces.Repositories;
-using Weblu.Application.Parameters;
 using Weblu.Application.Interfaces.Services.ServiceServices;
 using Weblu.Application.Common.Responses;
 using Weblu.Application.Common.Pagination;
+using Weblu.Application.Interfaces.Repositories.Services;
+using Weblu.Application.Dtos.Services.ServiceDtos;
+using Weblu.Application.Exceptions.CustomExceptions;
+using Weblu.Application.Parameters.Services;
+using Weblu.Application.Dtos.Services.ServiceDtos.ServiceImageDtos;
+using Weblu.Application.Interfaces.Repositories;
 
 namespace Weblu.Application.Services
 {
@@ -27,25 +28,22 @@ namespace Weblu.Application.Services
             _mapper = mapper;
             _serviceRepository = serviceRepository;
         }
-        public async Task<ServiceDetailDto> AddServiceAsync(AddServiceDto addServiceDto)
+        public async Task<ServiceDetailDto> CreateAsync(CreateServiceDto createServiceDto)
         {
-            Service newService = _mapper.Map<Service>(addServiceDto);
-            if (newService.IsActive)
-            {
-                newService.ActivatedAt = DateTimeOffset.Now;
-            }
+            Service newService = _mapper.Map<Service>(createServiceDto);
             _serviceRepository.Add(newService);
             await _unitOfWork.CommitAsync();
             ServiceDetailDto serviceDto = _mapper.Map<ServiceDetailDto>(newService);
             return serviceDto;
         }
-        public async Task DeleteServiceAsync(int serviceId)
+        public async Task DeleteAsync(int serviceId)
         {
             Service? service = await _serviceRepository.GetByIdAsync(serviceId) ?? throw new NotFoundException(ServiceErrorCodes.ServiceNotFound);
+            if (service.IsPublished) throw new ConflictException(ServiceErrorCodes.IsPublish);
             service.Delete();
             await _unitOfWork.CommitAsync();
         }
-        public async Task<PagedResponse<ServiceSummaryDto>> GetAllPagedServiceAsync(ServiceParameters serviceParameters)
+        public async Task<PagedResponse<ServiceSummaryDto>> GetAllPagedAsync(ServiceParameters serviceParameters)
         {
             PagedList<Service> services = await _serviceRepository.GetAllAsync(serviceParameters);
             List<ServiceSummaryDto> serviceSummaryDtos = _mapper.Map<List<ServiceSummaryDto>>(services);
@@ -54,13 +52,13 @@ namespace Weblu.Application.Services
             return pagedResponse;
         }
 
-        public async Task<List<ServiceSummaryDto>> GetAllServicesAsync(ServiceParameters serviceParameters)
+        public async Task<List<ServiceSummaryDto>> GetAllAsync(ServiceParameters serviceParameters)
         {
             IReadOnlyList<Service> services = await _serviceRepository.GetAllAsync(serviceParameters);
             List<ServiceSummaryDto> serviceSummaryDtos = _mapper.Map<List<ServiceSummaryDto>>(services);
             return serviceSummaryDtos;
         }
-        public async Task<ServiceDetailDto> GetServiceByIdAsync(int serviceId)
+        public async Task<ServiceDetailDto> GetByIdAsync(int serviceId)
         {
             Service? service = await _serviceRepository.GetByIdWithImagesAsync(serviceId) ?? throw new NotFoundException(ServiceErrorCodes.ServiceNotFound);
             List<ServiceImageDto> imageDtos = service.ServiceImages.Select(x => _mapper.Map<ServiceImageDto>(x)).ToList();
@@ -68,15 +66,35 @@ namespace Weblu.Application.Services
             serviceDto.Images = imageDtos;
             return serviceDto;
         }
-        public async Task<ServiceDetailDto> UpdateServiceAsync(int serviceId, UpdateServiceDto updateServiceDto)
+        public async Task<ServiceDetailDto> UpdateAsync(int serviceId, UpdateServiceDto updateServiceDto)
         {
             Service? service = await _serviceRepository.GetByIdAsync(serviceId) ?? throw new NotFoundException(ServiceErrorCodes.ServiceNotFound);
             service = _mapper.Map(updateServiceDto, service);
-            service.UpdateActivateStatus(updateServiceDto.IsActive);
+
+            service.MarkUpdated();
             _serviceRepository.Update(service);
             await _unitOfWork.CommitAsync();
+
             ServiceDetailDto serviceDto = _mapper.Map<ServiceDetailDto>(service);
             return serviceDto;
+        }
+
+        public async Task Publish(int serviceId)
+        {
+            Service? service = await _serviceRepository.GetByIdAsync(serviceId) ?? throw new NotFoundException(ServiceErrorCodes.ServiceNotFound);
+
+            service.Publish();
+
+            await _unitOfWork.CommitAsync();
+        }
+
+        public async Task Unpublish(int serviceId)
+        {
+            Service? service = await _serviceRepository.GetByIdAsync(serviceId) ?? throw new NotFoundException(ServiceErrorCodes.ServiceNotFound);
+
+            service.Unpublish();
+
+            await _unitOfWork.CommitAsync();
         }
     }
 }
